新增默认设备字典逻辑、app报警记录接口增加图片、修复场景保存bug

+++ application/src/main/java/org/thingsboard/server/controller/BoardSceneController.java
@@ -61,9 +61,10 @@ public class BoardSceneController extends BaseController {
         checkParameter("workshopId", workshopId);
         checkParameter("fileId", fileId);
         var fileInfo = this.fileService.getFileInfoByScopeAndEntityId(getTenantId(), FileScopeEnum.WORKSHOP_SCENE, toUUID(workshopId));
-        if (fileInfo != null)
+        if (fileInfo != null && !fileInfo.getId().equals(fileId)) {
             this.fileService.deleteFile(getTenantId(), fileInfo.getId());
-        this.fileService.updateFileScope(getTenantId(), toUUID(fileId), FileScopeEnum.WORKSHOP_SCENE, toUUID(workshopId));
+            this.fileService.updateFileScope(getTenantId(), toUUID(fileId), FileScopeEnum.WORKSHOP_SCENE, toUUID(workshopId));
+        }
     }
 
     /**
@@ -111,9 +112,10 @@ public class BoardSceneController extends BaseController {
         checkParameter("deviceId", deviceId);
         checkParameter("fileId", fileId);
         var fileInfo = this.fileService.getFileInfoByScopeAndEntityId(getTenantId(), FileScopeEnum.DEVICE_SCENE, toUUID(deviceId));
-        if (fileInfo != null)
+        if (fileInfo != null && !fileInfo.getId().equals(fileId)) {
             this.fileService.deleteFile(getTenantId(), fileInfo.getId());
-        this.fileService.updateFileScope(getTenantId(), toUUID(fileId), FileScopeEnum.DEVICE_SCENE, toUUID(deviceId));
+            this.fileService.updateFileScope(getTenantId(), toUUID(fileId), FileScopeEnum.DEVICE_SCENE, toUUID(deviceId));
+        }
     }
 
     /**
